+++ tcex/services/services.py
@@ -323,7 +323,7 @@ class Services(object):
             self.mqtt_client.loop_forever()
         except Exception as e:
             self.tcex.log.trace('error in listen_mqtt: {}'.format(e))
-            self.tcex.log.trace(traceback.format_exc())
+            self.tcex.log.error(traceback.format_exc())
 
     def listen_redis(self):
         """Listen for message coming from broker."""
@@ -751,9 +751,26 @@ class Services(object):
             method = message.get('method')
             params = message.get('queryParams')
             trigger_id = message.get('triggerId')
-            if self.webhook_event_callback(  # pylint: disable=not-callable
+            callback_response = self.webhook_event_callback(  # pylint: disable=not-callable
                 playbook, method, headers, params, body, config
-            ):
+            )
+            if isinstance(callback_response, dict):
+                # webhook responses are for providers that require a subscription req/resp.
+                webhook_event_response = {
+                    'sessionId': self.thread_name,  # session/context
+                    'requestKey': request_key,
+                    'command': 'WebhookEventResponse',
+                    'triggerId': trigger_id,
+                    'bodyVariable': 'response.body',
+                    'headers': callback_response.get('headers', []),
+                    'statusCode': callback_response.get('statusCode', 200),
+                }
+                # write response body to redis
+                playbook.create_string('response.body', callback_response.get('body'))
+
+                # publish the WebhookEventResponse message
+                self.publish(json.dumps(webhook_event_response))
+            elif isinstance(callback_response, bool) and callback_response:
                 self.increment_metric('hits')
                 self.fire_event_publish(trigger_id, self.thread_name, request_key)
             else: