+++ tcex/testing/test_case_job.py
@@ -51,7 +51,12 @@ class TestCaseJob(TestCase):
         if exit_code != 0:
             return exit_code
 
-        app.tcex.log.info('Exit Code: {}'.format(app.tcex.exit_code))
+        try:
+            # call exit for message_tc output, but don't exit
+            app.tcex.playbook.exit(msg=app.exit_message)
+        except SystemExit:
+            pass
+
         return self._exit(app.tcex.exit_code)
 
     def run_profile(self, profile):